fix: branch alignment
